move types to @types.ts files
