+++ src/cli/command/start/@types.ts
@@ -1,0 +1,3 @@
+export interface IStartCommandCliOptions {
+    'config-file': string;
+}
