+++ src/cli/command/start/builder.ts
@@ -5,9 +5,9 @@
 import * as fs from 'fs';
 import * as yargs from 'yargs';
 
-import { IStartCommandCliOptions } from './';
+import { IStartCommandCliOptions } from './@types';
 
-export default function builder(yargs: yargs.Argv<{}>): yargs.Argv<IStartCommandCliOptions> {
+export function builder(yargs: yargs.Argv<{}>): yargs.Argv<IStartCommandCliOptions> {
     return yargs.strict()
         .option('config-file', {
             coerce: function(path: string) {
