+++ src/cli/command/start/handler.ts
@@ -3,10 +3,10 @@
 // server.
 import { createServer } from '../../../server';
 
-import { IStartCommandCliOptions } from './';
+import { IStartCommandCliOptions } from './@types';
 import { parseConfigFile } from './config-file-parser';
 
-export default function handler(argv: IStartCommandCliOptions): void {
+export function handler(argv: IStartCommandCliOptions): void {
     const serverOptions = parseConfigFile(argv['config-file']),
         server = createServer(serverOptions);
