+++ src/cli/command/start/index.ts
@@ -1,11 +1,7 @@
 // A `yargs` CLI command module for the `raftjs`
 // server start command.
-import handler from './handler';
-import builder from './builder';
-
-export interface IStartCommandCliOptions {
-    'config-file': string;
-}
+import { handler } from './handler';
+import { builder } from './builder';
 
 export default {
     builder,
