+++ src/cluster/index.ts
@@ -1,2 +1,1 @@
 export * from './@types';
-export * from './tcp';
