+++ src/log/factory.ts
@@ -1,4 +1,5 @@
-import { ILog, ILogOptions, Log } from './log';
+import { ILog, ILogOptions } from './@types';
+import { Log } from './log';
 
 export function createLog(options?: ILogOptions): ILog {
     return new Log(options);
