+++ src/net/endpoint/@types.ts
@@ -1,0 +1,6 @@
+export interface IEndpoint {
+    equals: (endpoint: IEndpoint) => boolean;
+    host: string;
+    port: number;
+    toString: () => string;
+}
