+++ src/net/endpoint/index.ts
@@ -1,3 +1,3 @@
-export * from './endpoint';
+export * from './@types';
 export * from './factory';
 export * from './helpers';
