+++ src/server/election-timer/index.ts
@@ -1,4 +1,2 @@
-export * from './election-timer';
-export * from './election-timeout';
-export * from './election-timeout-chooser';
+export * from './@types';
 export * from './factory';
