+++ src/server/rpc/codec/@types.ts
@@ -1,0 +1,6 @@
+import { IRpcMessage } from '../message';
+
+export interface ICodec {
+    decode(data: Uint8Array): IRpcMessage;
+    encode(message: IRpcMessage): Uint8Array;
+}
