+++ src/server/rpc/codec/index.ts
@@ -1,7 +1,2 @@
-export { createFlatbuffersCodec } from './flatbuffers';
-import { IRpcMessage } from '../message';
-
-export interface ICodec {
-    decode(data: Uint8Array): IRpcMessage;
-    encode(message: IRpcMessage): Uint8Array;
-}
+export * from './@types';
+export * from './flatbuffers';
