+++ src/server/rpc/factory.ts
@@ -1,4 +1,5 @@
-import { IRpcService, IRpcServiceOptions, RpcService } from './service';
+import { IRpcService, IRpcServiceOptions } from './@types';
+import { RpcService } from './service';
 import { createFlatbuffersCodec } from './codec';
 import { createTcpTransport } from '../../transport';
