+++ src/server/rpc/index.ts
@@ -1,4 +1,4 @@
+export * from './@types';
 export * from './codec';
 export * from './factory';
 export * from './message';
-export * from './service';
