+++ src/server/rpc/receiver-registry/factory.ts
@@ -1,4 +1,5 @@
-import { IRpcReceiverRegistry, RpcReceiverRegistry } from './receiver-registry';
+import { IRpcReceiverRegistry } from './@types';
+import { RpcReceiverRegistry } from './receiver-registry';
 
 export function createRpcReceiverRegistry(): IRpcReceiverRegistry {
     return new RpcReceiverRegistry();
