+++ src/server/rpc/receiver-registry/index.ts
@@ -1,2 +1,2 @@
+export * from './@types';
 export * from './factory';
-export * from './receiver-registry';
