+++ src/server/rpc/receiver-registry/receiver-registry.ts
@@ -1,19 +1,7 @@
-import { RpcCallTypeMap, IRpcMessage, RpcProcedureTypeMap } from '../message';
+import { IRpcMessage, RpcCallTypeMap, RpcProcedureTypeMap } from '../message';
+import { IRpcReceiver } from '../@types';
+import { IRpcReceiverRegistry } from './@types';
 import { compilerError } from '../../../util/compiler-error';
-import { IRpcReceiver } from '../service';
-
-export interface IRpcReceiverRegistry {
-  add:    <P extends IRpcMessage['procedureType'], C extends IRpcMessage['callType']>(
-      receiver: IRpcReceiver<RpcProcedureTypeMap[P], RpcCallTypeMap[C]>
-  ) => void;
-  getAll: <P extends IRpcMessage['procedureType'], C extends IRpcMessage['callType']>(
-      procedureType: RpcProcedureTypeMap[P],
-      callType: RpcCallTypeMap[C]
-  ) => ReadonlySet<IRpcReceiver<RpcProcedureTypeMap[P], RpcCallTypeMap[C]>>;
-  remove: <P extends IRpcMessage['procedureType'], C extends IRpcMessage['callType']>(
-      receiver: IRpcReceiver<RpcProcedureTypeMap[P], RpcCallTypeMap[C]>
-  ) => void;
-}
 
 // A data structure that maintains an `RpcReceiver` set for each
 // kind of RPC message.
