+++ src/server/server.ts
@@ -18,46 +18,8 @@ import {
 import { IEndpoint, isEndpoint } from '../net/endpoint';
 import { IRpcEventListener, IRpcReceiver, IRpcService } from './rpc';
 import { IElectionTimer } from './election-timer';
+import { IServer, IServerOptions, ServerId } from './@types';
 import { IState, StateType, StateTransition, createState } from './state';
-
-export interface IServer {
-    readonly electionTimer: IElectionTimer;
-    readonly endpoint: IEndpoint;
-    readonly id: ServerId;
-    readonly log: ILog;
-    readonly logger: ILogger;
-    readonly peerApi: IRpcService;
-    getCommitIndex(): number;
-    getCluster(): ICluster;
-    getCurrentTerm(): number;
-    getLastApplied(): number;
-    getState(): IState;
-    getVotedFor(): ServerId;
-    onReceivePeerRpc<P extends IRpcMessage['procedureType'], C extends IRpcMessage['callType']>(
-        receiver: IRpcReceiver<P, C>
-    ): IRpcEventListener;
-    sendPeerRpc(message: IRpcMessage): Promise<Promise<void>[]>;
-    sendPeerRpc(endpoint: IEndpoint, message: IRpcMessage): Promise<Promise<void>[]>;
-    sendPeerRpc(endpoints: ReadonlyArray<IEndpoint>, message: IRpcMessage): Promise<Promise<void>[]>;
-    setCurrentTerm(newTerm: number): void;
-    setVotedFor(candidateId: ServerId): void;
-    start(): Promise<void>;
-    stop(): Promise<void>;
-    transitionTo(state: StateType | IState): void;
-}
-
-export type ServerId = string;
-
-export interface IServerOptions {
-    readonly cluster: ICluster;
-    readonly currentTerm: IDurableValue<number>;
-    readonly electionTimer: IElectionTimer;
-    readonly id: ServerId;
-    readonly log: ILog;
-    readonly logger: ILogger;
-    readonly peerApi: IRpcService;
-    readonly votedFor: IDurableValue<string>;
-}
 
 export class Server implements IServer {
     private readonly cluster: ICluster;
