--- src/server/state/noop.ts
@@ -1,3 +1,0 @@
-import { IState } from './';
-
-
