+++ src/storage/index.ts
@@ -1,2 +1,2 @@
-export * from './storage';
+export * from './@types';
 export * from './factory';
