+++ src/transport/connection-registry/factory.ts
@@ -1,4 +1,5 @@
-import { ConnectionRegistry, IConnectionRegistry } from './connection-registry';
+import { IConnectionRegistry } from './@types';
+import { ConnectionRegistry } from './connection-registry';
 
 export function createConnectionRegistry<T>(): IConnectionRegistry<T> {
     return new ConnectionRegistry();
