+++ src/transport/connection-registry/index.ts
@@ -1,2 +1,2 @@
-export * from './connection-registry';
+export * from './@types';
 export * from './factory';
