+++ src/transport/tcp/@types.ts
@@ -1,0 +1,7 @@
+import * as net from 'net';
+
+import { IConnectionRegistry } from '../connection-registry';
+
+export interface ITcpTransportOptions {
+    sockets?: IConnectionRegistry<net.Socket>;
+}
