+++ src/transport/tcp/factory.ts
@@ -1,5 +1,6 @@
 import { ITransport } from '../';
-import { ITcpTransportOptions, TcpTransport } from './tcp';
+import { ITcpTransportOptions } from './@types';
+import { TcpTransport } from './tcp';
 
 export function createTcpTransport(options: ITcpTransportOptions = {}): ITransport {
     return new TcpTransport(options);
