+++ src/transport/tcp/index.ts
@@ -1,2 +1,2 @@
+export * from './@types';
 export * from './factory';
-export * from './tcp';
